feat: implemented modular architecture with managers
