+++ tahrir_api/database/managers/__init__.py
@@ -1,0 +1,32 @@
+"""
+Database Managers Package
+
+Modular manager classes for Tahrir database operations.
+"""
+
+from .assertions_manager import AssertionManager
+from .authorizations_manager import AuthorizationManager
+from .badges_manager import BadgeManager
+from .base import BaseManager
+from .invitations_manager import InvitationManager
+from .issuers_manager import IssuerManager
+from .milestone_manager import MilestoneManager
+from .persons_manager import PersonManager
+from .ranking_manager import RankingManager
+from .series_manager import SeriesManager
+from .team_manager import TeamManager
+
+
+__all__ = [
+    "AssertionManager",
+    "AuthorizationManager",
+    "BadgeManager",
+    "BaseManager",
+    "InvitationManager",
+    "IssuerManager",
+    "MilestoneManager",
+    "PersonManager",
+    "RankingManager",
+    "SeriesManager",
+    "TeamManager",
+]
